+++ pkgs/unified_analytics/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Enhanced `LogFileStats` data to include information about flutter channel counts and tool counts
 - Added new method to suppress telemetry collection temporarily for current invocation via `analytics.suppressTelemetry()`
-- Added `SurveyHandler` feature to `Analytics` instance to fetch available surveys from remote endpoint to display to users
+- Added `SurveyHandler` feature to `Analytics` instance to fetch available surveys from remote endpoint to display to users along with functionality to dismiss them
 - Surveys will be disabled for any users that have been opted out
 
 ## 3.0.0